Refactor socket namespace creation again.
Now we setup the namespace with its options, and return
a factory interface for adding handlers and creation.

Add the fleshed out always-works auth layer to each socket handler,
so taht when a socket needs to be authorized it rejects requests
for non-authed users.  As well as provdes the user object in an efficent
manner.