Added a Socket schema, for telling sockets in a general sense messages that are critical.
This includes things like the 'UNAUTHORIZED' message, that
all sockets should listen for and immediately bump
their host to the login screens.

There is one remaining case for this that isn't handled yet:
 - when the server disconnects itself from the clients

Added socket disconnect handlers.

There is a distinction between a 'normal' and a 'last' handler,
simply meaning that some disconnect handlers are meant to run after
others.  In particular this is important for not unrolling the
AccountManager data before we can remove the person from all the
things the other handlers do.

Add more support in the socket manager for disconnects.

Working support for global chat.
Clients are now told when they join a global chat.

Removed the ChatManager's `onConnect` function, because
it doens't make sense for it to auto run.

Now we just expect the clients to ask us for a global chat,
and silently fail if they are already in one.