Add skeleton for game manager namespace at /game
The idea is that the game list will control putting people together,
and then when the host thinks the game is ready they can
signal the start of a game.

The players clients then all connect to the game server, and upon
connection will be entered into the proper game (after auth).

The game machine will control the game logic, and includes a new
implementation of a variable deck of cards.

When a card is delt the owner is marked. When played the cards state
should be updated to in play.  There may need to be additional state
data later on.

Once the game host has signaled a game to start, the pregame round
begins and betting starts.