Refactor the game start and ready/busy flow.
Now the gamelist namespace handles users ready and busy messages,
along with the host's start game command.

This now allows the game list manager to trigger a clone of the game
instance in the game manager.

At the point the host hits the 'ready' button on the game, the game
should (eventually) be removed from the game list and wholly owned
by the game server.

The game server should manage all leaving / rejoining, as well
as all game machine related events.